parser: avoid consing for buf literals.
* parser.y (buflit, buflit_items): Don't cons up a list of
bytes in buflit_items which are then assembled into a
buffer. Rather, the buflit_items rules construct and fill a
buffer object directly. The buflit rule then just has to
signal the end of the buffer literal to the lexer, and trim
the buffer to the actual size.  We will need this for
efficient loading of compiled files, in which the virtual
machine code is represented as a buffer literal.