compiler: move sys:bind-mac-error to error.tl.
The sys:bind-mac-error function is a run-time support function
for the compiler-generated code for destructuring.  But the
compiler.tl source file is a bad place for it.  It means that
compiled modules which need sys:bind-mac-error auto-load the
entire compiler. That easily creates an irresolveable cycle,
because the compiler is on top of the foodchain.

No, sys:bind-mac-error belongs in error.tl; the light-weight
module of low-dependency run-time support routines for error
reporting.

* lisplib.c (error_set_entries): Add sys:bind-mac-error symbol
here.
(compiler_set_entries): Remove it from here.

* share/txr/stdlib/compiler.tl (sys:bind-mac-error): Function
removed.

* share/txr/stdlib/error.tl (sys:bind-mac-error): Function
moved here.