configure: link files rather than dirs.
* configure (lndir): New function; imitation of same-named
utility from the X11 distribution. Using lndir, instead of
creating, in the build directory, symlinks for the tests and
shared directories pointing back to the source tree,
we mirror the directory structure of these directories and
populate it with individual file symlinks. Motivation:
this will allow us to build .tlo object files in these
directories side by side with .tl files, without touching
the source tree.