compiler: fix overestimated register count.
The compiler internally allocates registers that are never
inserted into code (but correctly recycled). Thus the treg
bump counter overestimates the actual number of registers that
a VM requires; but that's how we use it! That causes wasted
stack space.

What we can do to get a tightly clamped register count
is simply to let the assembler report the highest-numbered
treg operand that it finds in the code.

(In the future we will need something better to optimize the
register allocation. Why: because closures have their own VM
instance with its own registers. If an overall VM uses
registers t00 through t09, but a certain closure uses
only t02 through t06, that closure's frame only needs to
allocate seven registers.)

* share/txr/stdlib/asm.tl (assembler): New slot, max-treg.
(assembler parse-args): Whenever a register operand is parsed,
if it is a t register, update max-treg.

* share/txr/stdlib/compiler.tl (usr:compile-toplevel): When
creating the vm-desc, determine the nreg parameter from the
assembler's max-treg value, rather than from the compiler's
treg allocation bump counter.