awk/struct: fix breakage under interpreted mode.
References to rslot must be usr:rslot because there is a
sys:rslot. Some difference in loading order causes this not to
be a problem when runing compiled "make tests", but if the
.tlo files are removed, the awk test case fails.

* share/txr/stdlib/awk.tl (sys:awk-mac-let): Replace
rslot with usr:rslot.

* share/txr/stdlib/struct.tl (rslot): Replace rslot
symbol token with usr:rslot in definitions of rslot macro and
rslot place macro.