Makefile: bugfix: shell quoting issue in macro.
* Makefile (SH): The $(1) macro argument is substituted into
a shell '...' quote. It may contain single quotes itself,
so these have to be mapped to the famous '\'' sequence.
Currently, only one SH call has '...' syntax in it: one
in the y.tab.c rule which calls sed -e '/yyparse/d'.
That works anyway, and the quotes could be removed from
it; but let's fix the SH macro.