build: remove dubious license header from y.tab.c.
The GNU Bison parser has a GPL header in the parser skeleton
which gets copied to the generated parser. The GPL asserts
that the entire file falls under the GPL (but as a special
exception, it may be used in any manner whatsoever in a
program, as long as it isn't used as the skeleton of a parser
generator).  The GPL assertion is false; the file is not
entirely derived from the skeleton but contains code directly
copied from parser.y, as well as tables generated from
parser.y's original grammar. Since we are definitely not using
this as a parser skeleton for a parser generator and are thus
allowed to do whatever we want such as compiling this into
proprietary object files, I'm removing the copyright header as
a build step. You never know; someone could package up a built
directory of TXR containing a Bison-generated y.tab.c, and
then it looks like there is a GPL-ed file in the project.

* Makefile (CM): New variable, needed for interpolating
literal commas into macro calls.
(BS_LIC_FROM, BS_LIC_TO): New variables.
(y.tab.c): Filter y.tab.c to remove GPL comment.