compiler: bugfix: miscompiled unwind-protect.
* share/txr/stdlib/compiler.tl (compiler comp-unwind-protect):
The protected code must terminate with an end instruction. It
cannot just branch to the skip label because then we keep
executing the following unprotected code still under the
unwind protect. How it works is that the protected code ends
with end. Then the cleanup fragment at the cleanup offset is
executed, also terminating with an end (that one we have in
place).  We don't need the skip label because the instruction
pointer advances after executing the end instruction after the
cleanup fragment.