build: fix failure due to .tlo compilation order.
On newer GNU Make versions, $(wildcard ...) doesn't sort
its results, and so we compile .tlo files in an
arbitrary order. For the most part, this is fine.
However, there is a complicated circular dependency involving
the error.tl file requiring it to be compiled first.

* Makefile (STLIB_EARLY_PATS, STDLIB_EARLY_TLOS,
STDLIB_LATE_TLOS): New variables. Use these
to express a dependency which causes error.tlo
to be built first.